Add traceable ID to traceable directive output
This change adds IDs to the output of traceable directives, so that
their HTML output contains an "id" attribute and can be linked to using
the "url#id" format. This is necessary when linking to a traceable on a
long HTML page for browsers to know which part of that page to display
initially.

Resolves #2.